Replace treatments table with view
Add the treatment name to the "happens-before" relation between tasks in
`taskDependencies`, since the same task can be in different positions in
differrent treatments.

Then, replace the `treatments` table with a view that uses the adjancency list
in `taskDependencies` to calculate the order of tasks in a treatment with
the given name.

Note that two tasks in a treatment can have the same `taskPrio` order if
there is no "happens-before" relation between them.  The model does not
currently support this case.